Bugfix: User session not available on conversation share action
Issue:
user session not available on conversation share

Root cause:
In current implementation the share logic (index.js) will be executed in content of chat gpt website. Because of the context, the user session of the extension is not available on share to server.

Fix:
Extension part:
1) Added background.js
2) Extended manifest.json file by background.js for background handling
3) Moved share logic to background.js
4) Implemented communication between extension background process and chat gpt website with share button (index.js) to send share command to background task

Share gpt server part:
In index.ts:
1) Added session validation, share request will be rejected in case no valid session exists
2) Renamed setRandomKey to saveConversationAndGetConversationHash
3) Updated prism data to store userId as creatorId